+++ Converter/MapManager.cs
@@ -27,7 +27,7 @@ public static class MapManager
     {
         try
         {
-            var file = await File.ReadAllTextAsync(Settings.Instance.inputGeojsonPath);
+            var file = await File.ReadAllTextAsync(Settings.Instance.InputGeojsonPath);
             var geomap = JsonSerializer.Deserialize(file, GeoMapJsonContext.Default.GeoMap);
             return geomap;
         }
@@ -55,7 +55,7 @@ public static class MapManager
     {
         try
         {
-            var file = await File.ReadAllTextAsync(Settings.Instance.inputJsonPath);
+            var file = await File.ReadAllTextAsync(Settings.Instance.InputJsonPath);
             var jsonmap = JsonSerializer.Deserialize(file, JsonMapJsonContext.Default.JsonMap);
             return jsonmap;
         }
@@ -413,7 +413,7 @@ public static class MapManager
             }
 
             cellsMap.Draw(drawables);
-            var path = Helper.GetPath(SettingsManager.OutputDirectory, "map_data", "provinces.png");
+            var path = Helper.GetPath(Settings.OutputDirectory, "map_data", "provinces.png");
             Directory.CreateDirectory(Path.GetDirectoryName(path));
             await cellsMap.WriteAsync(path);
         }
@@ -471,7 +471,7 @@ public static class MapManager
             }
 
             cellsMap.Draw(drawables);
-            var path = Helper.GetPath(SettingsManager.OutputDirectory, "map_data", "heightmap.png");
+            var path = Helper.GetPath(Settings.OutputDirectory, "map_data", "heightmap.png");
             Directory.CreateDirectory(Path.GetDirectoryName(path));
             await cellsMap.WriteAsync(path);
 
@@ -521,7 +521,7 @@ public static class MapManager
             }
 
             cellsMap.Draw(drawables);
-            var path = Helper.GetPath(SettingsManager.OutputDirectory, "map_data", "rivers.png");
+            var path = Helper.GetPath(Settings.OutputDirectory, "map_data", "rivers.png");
             Directory.CreateDirectory(Path.GetDirectoryName(path));
 
             cellsMap.Settings.SetDefine("png:color-type", "1");
@@ -560,7 +560,7 @@ public static class MapManager
     public static async Task WriteDefinition(Map map)
     {
         var lines = map.Provinces.Select((n, i) => $"{i};{n.Color.R};{n.Color.G};{n.Color.B};{n.Name};x;");
-        var path = Helper.GetPath(SettingsManager.OutputDirectory, "map_data", "definition.csv");
+        var path = Helper.GetPath(Settings.OutputDirectory, "map_data", "definition.csv");
         Directory.CreateDirectory(Path.GetDirectoryName(path));
         await File.WriteAllLinesAsync(path, lines);
     }
@@ -594,7 +594,7 @@ $@"game_object_locator={{
 {string.Join("\n", lines)}
     }}
 }}";
-            var path = Helper.GetPath(SettingsManager.OutputDirectory, "gfx", "map", "map_object_data", "building_locators.txt");
+            var path = Helper.GetPath(Settings.OutputDirectory, "gfx", "map", "map_object_data", "building_locators.txt");
             Directory.CreateDirectory(Path.GetDirectoryName(path));
             await File.WriteAllTextAsync(path, file);
         }
@@ -633,7 +633,7 @@ $@"game_object_locator={{
 {string.Join("\n", lines)}
     }}
 }}";
-            var path = Helper.GetPath(SettingsManager.OutputDirectory, "gfx", "map", "map_object_data", "siege_locators.txt");
+            var path = Helper.GetPath(Settings.OutputDirectory, "gfx", "map", "map_object_data", "siege_locators.txt");
             Directory.CreateDirectory(Path.GetDirectoryName(path));
             await File.WriteAllTextAsync(path, file);
         }
@@ -672,7 +672,7 @@ $@"game_object_locator={{
 {string.Join("\n", lines)}
     }}
 }}";
-            var path = Helper.GetPath(SettingsManager.OutputDirectory, "gfx", "map", "map_object_data", "combat_locators.txt");
+            var path = Helper.GetPath(Settings.OutputDirectory, "gfx", "map", "map_object_data", "combat_locators.txt");
             Directory.CreateDirectory(Path.GetDirectoryName(path));
             await File.WriteAllTextAsync(path, file);
         }
@@ -725,7 +725,7 @@ $@"game_object_locator={{
 {string.Join("\n", lines)}
     }}
 }}";
-            var path = Helper.GetPath(SettingsManager.OutputDirectory, "gfx", "map", "map_object_data", "player_stack_locators.txt");
+            var path = Helper.GetPath(Settings.OutputDirectory, "gfx", "map", "map_object_data", "player_stack_locators.txt");
             Directory.CreateDirectory(Path.GetDirectoryName(path));
             await File.WriteAllTextAsync(path, file);
         }
@@ -794,7 +794,7 @@ sea_zones = LIST {{ {string.Join(" ", waterProvinces)} }}
 # They are probably not visible anywhere on the map, so feel free to reuse them (after double checking that they are actually missing).
 ";
 
-        var path = Helper.GetPath(SettingsManager.OutputDirectory, "map_data", "default.map");
+        var path = Helper.GetPath(Settings.OutputDirectory, "map_data", "default.map");
         Directory.CreateDirectory(Path.GetDirectoryName(path));
         await File.WriteAllTextAsync(path, file);
     }
@@ -823,7 +823,7 @@ sea_zones = LIST {{ {string.Join(" ", waterProvinces)} }}
 
             var file = $@"default=plains
 {string.Join("\n", provinceBiomes)}";
-            var path = Helper.GetPath(SettingsManager.OutputDirectory, "common", "province_terrain", "00_province_terrain.txt");
+            var path = Helper.GetPath(Settings.OutputDirectory, "common", "province_terrain", "00_province_terrain.txt");
             Directory.CreateDirectory(Path.GetDirectoryName(path));
             await File.WriteAllTextAsync(path, file);
         }
@@ -852,7 +852,7 @@ sea_zones = LIST {{ {string.Join(" ", waterProvinces)} }}
 
             cellsMap.Draw(drawables);
 
-            var path = Helper.GetPath(SettingsManager.OutputDirectory, "gfx", "map", "terrain", $"{filename}.png");
+            var path = Helper.GetPath(Settings.OutputDirectory, "gfx", "map", "terrain", $"{filename}.png");
             Directory.CreateDirectory(Path.GetDirectoryName(path));
             await cellsMap.WriteAsync(path, MagickFormat.Png00);
 
@@ -954,7 +954,7 @@ sea_zones = LIST {{ {string.Join(" ", waterProvinces)} }}
 
     public static async Task WriteGraphics()
     {
-        var path = Helper.GetPath(SettingsManager.OutputDirectory, "common", "defines", "graphic", "00_graphics.txt");
+        var path = Helper.GetPath(Settings.OutputDirectory, "common", "defines", "graphic", "00_graphics.txt");
         Directory.CreateDirectory(Path.GetDirectoryName(path));
         File.Copy(Helper.GetPath(SettingsManager.ExecutablePath, "00_graphics.txt"), path, true);
     }
@@ -1098,7 +1098,7 @@ sea_zones = LIST {{ {string.Join(" ", waterProvinces)} }}
                     .ToArray();
 
                 var file = string.Join('\n', baronies);
-                var path = Helper.GetPath(SettingsManager.OutputDirectory, "history", "provinces", $"k_{kingdom.id}.txt");
+                var path = Helper.GetPath(Settings.OutputDirectory, "history", "provinces", $"k_{kingdom.id}.txt");
                 Directory.CreateDirectory(Path.GetDirectoryName(path));
                 await File.WriteAllTextAsync(path, file);
             }
@@ -1106,8 +1106,8 @@ sea_zones = LIST {{ {string.Join(" ", waterProvinces)} }}
 
         // Make original culture history empty
         // Otherwise it will override newly created culture
-        var originalProvincesPath = Helper.GetPath(map.Settings.ck3Directory, "history", "provinces");
-        var provincesPath = Helper.GetPath(SettingsManager.OutputDirectory, "history", "provinces");
+        var originalProvincesPath = Helper.GetPath(map.Settings.Ck3Directory, "history", "provinces");
+        var provincesPath = Helper.GetPath(Settings.OutputDirectory, "history", "provinces");
         foreach (var p in Directory.EnumerateFiles(originalProvincesPath))
         {
             File.WriteAllText(provincesPath + Path.GetFileName(p), "");
@@ -1118,15 +1118,15 @@ sea_zones = LIST {{ {string.Join(" ", waterProvinces)} }}
         try
         {
             // Delete religion file from Total conversion sandbox mod.
-            File.Delete(Helper.GetPath(SettingsManager.OutputDirectory, "common", "religion", "religions", "01_vanilla.txt"));
+            File.Delete(Helper.GetPath(Settings.OutputDirectory, "common", "religion", "religions", "01_vanilla.txt"));
         }
         catch
         {
             // Do nothing.
         }
 
-        var religionsPath = Helper.GetPath(map.Settings.ck3Directory, "common", "religion", "religions");
-        FileSystem.CopyDirectory(religionsPath, Helper.GetPath(SettingsManager.OutputDirectory, "common", "religion", "religions"), true);
+        var religionsPath = Helper.GetPath(map.Settings.Ck3Directory, "common", "religion", "religions");
+        FileSystem.CopyDirectory(religionsPath, Helper.GetPath(Settings.OutputDirectory, "common", "religion", "religions"), true);
     }
     // Maps original holy sites to newly created provinces.
     public static async Task WriteHolySites(Map map, string[] pickedFaiths)
@@ -1163,7 +1163,7 @@ sea_zones = LIST {{ {string.Join(" ", waterProvinces)} }}
         });
 
         var file = string.Join('\n', mappedHolySites);
-        var path = Helper.GetPath(SettingsManager.OutputDirectory, "common", "religion", "holy_sites", "00_holy_sites.txt");
+        var path = Helper.GetPath(Settings.OutputDirectory, "common", "religion", "holy_sites", "00_holy_sites.txt");
         Directory.CreateDirectory(Path.GetDirectoryName(path));
         await File.WriteAllTextAsync(path, file);
     }